+++ packages/probes/src/sectum/probes/ikea_extraction/probe.py
@@ -22,8 +22,8 @@ class IkeaExtractionProbe(DetectingProbe):
     id = "ikea-extraction"
     name = "IKEA-style implicit benign extraction"
     owasp_llm = "LLM08:2025"
-    atlas_techniques: tuple[str, ...] = ()
-    nist_rmf: tuple[str, ...] = ()
+    atlas_techniques: tuple[str, ...] = ("AML.T0024", "AML.T0057")
+    nist_rmf: tuple[str, ...] = ("MEASURE 2.7",)
     surfaces: tuple[Surface, ...] = (Surface.VECTOR_DB,)
     requires_adapters: tuple[str, ...] = ()
 
@@ -62,4 +62,7 @@ class IkeaExtractionProbe(DetectingProbe):
             observation.surface,
             probe_id=self.id,
             observed_user=step.actor_user_id,
+            owasp_llm=self.owasp_llm,
+            atlas=self.atlas_techniques,
+            nist=self.nist_rmf,
         )